Title: 합과 평균, Time: 6ms, Memory: 0MB, Status: Wrong Answer - Codetree

+++ 250522/합과 평균/sum-and-avg.cpp
@@ -5,6 +5,6 @@ int main() {
     int A,B;
     cin >> A >> B;
 
-    cout << A+B << " " << (A+B)/2;
+    cout << A+B << " " << (double)(A+B)/2;
     return 0;
 }
